requirements.txt
pip>=20.0
django
djangorestframework
psycopg2-binary
django-environ
drf-spectacular
django-cors-headers
django-redis
